+++ lib/shopify-cli/commands/switch.rb
@@ -4,14 +4,22 @@ module ShopifyCli
   module Commands
     class Switch < ShopifyCli::Command
       options do |parser, flags|
-        parser.on("--shop=SHOP") do |shop|
-          flags[:shop] = shop
-        end
+        parser.on("--store=STORE") { |url| flags[:shop] = url }
+        # backwards compatibility allow 'shop' for now
+        parser.on("--shop=SHOP") { |url| flags[:shop] = url }
       end
 
       def call(*)
+        if Shopifolk.acting_as_shopify_organization?
+          @ctx.puts(@ctx.message("core.switch.disabled_as_shopify_org"))
+          return
+        end
+
         shop = if options.flags[:shop]
           Login.validate_shop(options.flags[:shop])
+        elsif (org_id = DB.get(:organization_id))
+          res = ShopifyCli::Tasks::SelectOrgAndShop.call(@ctx, organization_id: org_id)
+          res[:shop_domain]
         else
           AdminAPI.get_shop_or_abort(@ctx)
           res = ShopifyCli::Tasks::SelectOrgAndShop.call(@ctx)